requirement.txt
alembic==1.3.2
asgiref==3.2.3
certifi==2019.11.28
cycler==0.10.0
dataloader==2.0
dataset==1.1.2
Django==3.0.2
f==0.0.1
image==1.5.27
kiwisolver==1.1.0
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.18.1
pandas==0.25.3
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.12
sqlparse==0.3.0
torch==1.4.0
torchvision==0.5.0
transforms==0.1
